examples/llms/search_grounding_stub.py
"""Search grounding demonstration stub.

This script sketches a grounding workflow combining query planning,
web search (placeholder), evidence ranking, and answer synthesis.
Replace placeholder functions with actual implementations / APIs.
"""
from dataclasses import dataclass
from typing import List, Dict
import re

# If google-genai is configured, you could import and initialize a client.
try:
    import google.genai as genai  # type: ignore
    CLIENT_AVAILABLE = True
except Exception:
    CLIENT_AVAILABLE = False

@dataclass
class Evidence:
    source: str
    snippet: str
    score: float


def plan_queries(question: str) -> List[str]:
    # Simple heuristic: split into keywords sets
    base = re.sub(r"[^a-zA-Z0-9 ]", "", question).lower().split()
    unique = list(dict.fromkeys(base))
    return [" ".join(unique), question]


def web_search(query: str) -> List[Dict]:
    # Placeholder search results; integrate real API here
    return [
        {"url": f"https://example.com/{i}", "title": f"Result {i} for {query}", "snippet": f"Snippet about {query} #{i}"}
        for i in range(1, 4)
    ]


def collect_evidence(queries: List[str]) -> List[Evidence]:
    ev: List[Evidence] = []
    for q in queries:
        for r in web_search(q):
            ev.append(Evidence(source=r["url"], snippet=r["snippet"], score=len(r["snippet"])) )
    # naive ranking by snippet length
    return sorted(ev, key=lambda e: e.score, reverse=True)[:5]


def synthesize_answer(question: str, evidence: List[Evidence]) -> Dict:
    joined = " \n".join([f"[{i+1}] {e.snippet} (source: {e.source})" for i, e in enumerate(evidence)])
    answer = (
        f"Question: {question}\n\n" +
        "Evidence:\n" + joined + "\n\n" +
        "Draft Answer: Based on the collected sources, summarize key points here."
    )
    return {"answer": answer, "citations": [e.source for e in evidence]}


def main():
    question = "What are recent trends in open-source LLM fine-tuning?"
    queries = plan_queries(question)
    evidence = collect_evidence(queries)
    result = synthesize_answer(question, evidence)
    print(result["answer"])
    print("Citations:", result["citations"])


if __name__ == "__main__":
    main()
