examples/llms/requirements.txt
transformers
datasets
peft
accelerate
scikit-learn
bitsandbytes
google-genai